requirements.txt
django
pillow
six
psycopg2
django-allauth
djangorestframework
daphne
channels
celery
openai
drf-yasg
coverage
asgiref
pytz
black